First version of the Dereferencibility of Forward/Backward links metrics and the respective unit tests. This implementation follows the formula:
(# of outlinks of the resource that are contained into it) / (Total # of subject URIs in the resource)

This version is based on the fact that in our case, we have access to a resource at a time, that is, we don't have access to the whole dataset and therefore, we cannot compute “the number of outlinks of the resource in the whole dataset” as required by Hogan et al. in "An empirical survey of Linked Data conformance".